// src/service/file/schemas/studentEvaluationIndicatorResult.ts
/**
 * Generated by orval v7.4.1 🍺
 * Do not edit manually.
 * File API
 * isd / kinder board api doc
 * OpenAPI spec version: v1.1.0
 */
import type { StudentEvaluationIndicatorResultCourse } from './studentEvaluationIndicatorResultCourse';

export interface StudentEvaluationIndicatorResult {
  /** 지표의 코드 */
  indicatorCode: string;
  /** 지표가 해당하는 교육 과정 */
  course: StudentEvaluationIndicatorResultCourse;
  /** 지표가 해당하는 영역 */
  evaluationDomainCode: string;
  /** 정렬 순서. 0 부터 */
  indicatorOrder: number;
  /** 지표 이름 */
  title?: string;
  /**
   * 지표 설명
   * @nullable
   */
  description: string | null;
  /**
   * 지표 중분류 설명
   * @nullable
   */
  referenceData?: string | null;
  /** 뎁스 */
  depth: number;
  /** 생성일 */
  createdAt: string;
  /** 수정일 */
  modifiedAt: string;
  /** 하위 항목들 */
  subIndicators?: StudentEvaluationIndicatorResult[];
}