// src/service/file/schemas/studentRecordAddRequest.ts
/**
 * Generated by orval v7.4.1 🍺
 * Do not edit manually.
 * File API
 * isd / kinder board api doc
 * OpenAPI spec version: v1.1.0
 */
import type { StudentRecordAddRequestCourse } from './studentRecordAddRequestCourse';
import type { StudentRecordEvaluationIndicatorScoreForAdd } from './studentRecordEvaluationIndicatorScoreForAdd';
import type { StudentRecordAttachedPhotoForAdd } from './studentRecordAttachedPhotoForAdd';
import type { StudentRecordCommentForAdd } from './studentRecordCommentForAdd';

export interface StudentRecordAddRequest {
  /**
   * 작성하는 선생님의 계정 아이디
   * @minimum 0
   */
  creatorAccountId: number;
  /**
   * 작성하는 선생님의 프로필 아이디
   * @minimum 0
   */
  creatorProfileId: number;
  /**
   * 아이가 속한 반의 아이디
   * @minimum 0
   */
  educationalClassId: number;
  /**
   * 아이의 아이디
   * @minimum 0
   */
  studentId: number;
  /**
   * 아이의 이름
   * @minLength 1
   * @maxLength 45
   */
  studentName: string;
  /** 아이의 나이 */
  studentAge: number;
  /**
   * 아이가 현재 해당하는 교육 과정
   * @nullable
   */
  course?: StudentRecordAddRequestCourse;
  /**
   * 방사형 그래프의 영역과 하위 항목들의 평균 점수
   * @maxLength 30
   * @nullable
   */
  summaryScores?: string | null;
  /**
   * 대표사진 url
   * @maxLength 1000
   * @nullable
   */
  thumbUrl?: string | null;
  /**
   * 한줄평 이미지
   * @maxLength 1000
   * @nullable
   */
  typeImageUrl?: string | null;
  /**
   * 한줄평 내용
   * @maxLength 200
   * @nullable
   */
  typeDescription?: string | null;
  /**
   * 그래프에 들어간 항목에 대한 점수들
   * @nullable
   */
  indicatorScores?: StudentRecordEvaluationIndicatorScoreForAdd[] | null;
  /**
   * 첨부된 사진들
   * @nullable
   */
  attachedPhotos?: StudentRecordAttachedPhotoForAdd[] | null;
  /**
   * 관찰 평가 결과 영역별 내용들
   * @nullable
   */
  observeComments?: StudentRecordCommentForAdd[] | null;
  /** ai가 생성한 교사지원 내용 */
  teacherSupport?: string;
  /** 교사지원 내용을 사용자가 직접 수정 했는지? */
  teacherSupportUserEdited?: boolean;
  /** ai가 생성한 학부모 지원 내용 */
  parentSupport?: string;
  /** 학부모 지원 내용을 사용자가 직접 수정 했는지? */
  parentSupportUserEdited?: boolean;
  /** ai가 생성한 교사 총평 내용 */
  teacherComment?: string;
  /** 교사 총평 내용을 사용자가 직접 수정 했는지? */
  teacherCommentUserEdited?: boolean;
  /** 관찰 요약 내용 */
  observeSummary?: string;
  /** 관찰 요약 내용을 사용자가 직접 수정 했는지? */
  observeSummaryUserEdited?: boolean;
  /**
   * 작성하는 선생님의 ip
   * @minLength 1
   * @maxLength 40
   */
  creatorIp: string;
}