// src/service/file/schemas/studentRecordUpdateRequest.ts
/**
 * Generated by orval v7.4.1 🍺
 * Do not edit manually.
 * File API
 * isd / kinder board api doc
 * OpenAPI spec version: v1.1.0
 */
import type { StudentRecordEvaluationIndicatorScoreForUpdate } from './studentRecordEvaluationIndicatorScoreForUpdate';
import type { StudentRecordAttachedPhotoForUpdate } from './studentRecordAttachedPhotoForUpdate';
import type { StudentRecordCommentForUpdate } from './studentRecordCommentForUpdate';

export interface StudentRecordUpdateRequest {
  /**
   * 관찰기록의 id (StudentRecord.id)
   * @minimum 0
   */
  id: number;
  /**
   * 작성하는 선생님의 프로필 아이디
   * @minimum 0
   */
  creatorProfileId: number;
  /**
   * 방사형 그래프의 영역과 하위 항목들의 평균 점수
   * @maxLength 30
   * @nullable
   */
  summaryScores?: string | null;
  /**
   * 대표사진 url
   * @maxLength 1000
   * @nullable
   */
  thumbUrl?: string | null;
  /**
   * 한줄평 이미지
   * @maxLength 1000
   * @nullable
   */
  typeImageUrl?: string | null;
  /**
   * 한줄평 내용
   * @maxLength 200
   * @nullable
   */
  typeDescription?: string | null;
  /**
   * 그래프에 들어간 항목에 대한 점수들
   * @nullable
   */
  indicatorScores?: StudentRecordEvaluationIndicatorScoreForUpdate[] | null;
  /**
   * 첨부된 사진들
   * @nullable
   */
  attachedPhotos?: StudentRecordAttachedPhotoForUpdate[] | null;
  /**
   * 관찰 평가 결과 영역별 내용들
   * @nullable
   */
  observeComments?: StudentRecordCommentForUpdate[] | null;
  /** 교사지원 내용 */
  teacherSupport?: string;
  /** 교사지원 내용을 사용자가 직접 수정 했는지? */
  teacherSupportUserEdited?: boolean;
  /** 학부모 지원 내용 */
  parentSupport?: string;
  /** 학부모 지원 내용을 사용자가 직접 수정 했는지? */
  parentSupportUserEdited?: boolean;
  /** 교사 총평 내용 */
  teacherComment?: string;
  /** 교사 총평 내용을 사용자가 직접 수정 했는지? */
  teacherCommentUserEdited?: boolean;
  /** 관찰 요약 내용 */
  observeSummary?: string;
  /** 관찰 요약 내용을 사용자가 직접 수정 했는지? */
  observeSummaryUserEdited?: boolean;
  /**
   * 작성하는 선생님의 ip
   * @minLength 1
   * @maxLength 40
   */
  creatorIp: string;
}